pyproject.toml
[tool.poetry]
name = "olympus"
version = "0.1.0"
description = "NB Secure-ish Hacker Entrance Device"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"RPi.GPIO" = {version = "^0.7.1", platform = "linux"}
spidev = {version = "^3.6", platform = "linux"}
mfrc522 = {version = "^0.0.7", platform = "linux"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"